scripts/requirements.txt
PyYAML>=6.0
Pillow>=10.0
google-genai>=1.0
